create createdOn on create

+++ DTOs/Recommendations/ReadRecommendation.cs
@@ -16,7 +16,7 @@ namespace DTOs.Recommendations
         public string Facebook { get; set; }
         public string OtherLink { get; set; }
         public string Photo { get; set; }
-        public int CreatedOn { get; set; }
+        public long CreatedOn { get; set; }
         public string CityDId { get; set; }
         public string[] Tags { get; set; }
         public string FromUserDId { get; set; }
@@ -24,7 +24,7 @@ namespace DTOs.Recommendations
 
         public ReadRecommendation(string dId, string placeName, string title, string text,
             string address, string maps, string website, string instagram, string facebook,
-            string otherLink, string photo, int createdOn, string cityDId, string[] tags,
+            string otherLink, string photo, long createdOn, string cityDId, string[] tags,
             string fromUserDId, string toUserDId)
         {
             DId = dId;
